Moved models into folder
