+++ src/app/shop/models/product.ts
