+++ src/app/shop/product-type-select/product-type-select.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input } from '@angular/core';
-import { ProductType } from '../services/product-type';
+import { ProductType } from '../models/product-type';
 
-import { ProductTypeModel } from '../services/product-type-model';
+import { ProductTypeModel } from '../models/product-type-model';
 
 @Component({
   selector: 'bp-product-type-select',
