+++ src/app/shop/resolves/types.resolve.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Resolve } from '@angular/router';
 
-import { ProductTypeModel } from '../services/product-type-model';
+import { ProductTypeModel } from '../models/product-type-model';
 import { ShopResolvesModule } from './shop-resolves.module';
 import { ShopService } from '../services/shop.service';
